refactor: adjust tests for 1e18

+++ src/yield-exposed-tokens/yeETH/YeETH.sol
@@ -17,7 +17,7 @@ contract YeETH is YieldExposedToken {
         string calldata name_,
         string calldata symbol_,
         address underlyingToken_,
-        uint8 minimumReservePercentage_,
+        uint256 minimumReservePercentage_,
         address yieldVault_,
         address yieldRecipient_,
         address lxlyBridge_,
